feat: implement EUTrialParser for parsing EU CTIS RSS feed and integrate with feedreader_trials command
